+++ ksa_compliance/__init__.py
@@ -1,3 +1,8 @@
+import logging
+
+from frappe.utils.logger import get_logger
+
+logger = get_logger('zatca')
+logger.setLevel(logging.INFO)
 
 __version__ = '0.0.1'
-