+++ ksa_compliance/generate_xml.py
@@ -1,5 +1,8 @@
-import frappe
 import json
+
+import frappe
+
+from ksa_compliance.invoice import InvoiceType
 
 
 @frappe.whitelist(allow_guest=True)
@@ -21,17 +24,19 @@ def generate_xml(input_data: dict = None):
                  "business_settings": data.get("business_settings")},
         is_path=True
     )
+
+    # TODO: This is not good; it works because the template doesn't contain any entities currently. But it'll break if
+    #  any are added
     invoice_xml = invoice_xml.replace("&", "&amp;")
     invoice_xml = invoice_xml.replace("\n", "")
 
     return invoice_xml
 
 
-def generate_xml_file(data, invoice_type: str = "Simplified"):
+def generate_xml_file(data: dict, invoice_type: InvoiceType = "Simplified"):
     """
     For Generating cycle
     """
-
     template = "simplified_e_invoice.xml" if invoice_type.lower() == "simplified" else "standard_e_invoice.xml"
 
     # render XML Template
@@ -44,28 +49,13 @@ def generate_xml_file(data, invoice_type: str = "Simplified"):
             "business_settings": data.get("business_settings")},
         is_path=True
     )
+    # TODO: Fix this
     invoice_xml = invoice_xml.replace("&", "&amp;")
     invoice_xml = invoice_xml.replace("\n", "")
-    xml_filename = generate_einvoice_xml_fielname(data['business_settings'], data["invoice"])
-    file = frappe.get_doc(
-        {
-            "doctype": "File",
-            "file_name": xml_filename,
-            "attached_to_doctype": "Sales Invoice",
-            "attached_to_name": data["invoice"]["id"],
-            "content": invoice_xml,
-        }
-    )
-    file.insert()
     return invoice_xml
 
 
-def generate_einvoice_xml_fielname(business_settings, invoice):
-    vat_registration_number = business_settings["company_id"]
-    invoice_date = invoice["issue_date"].replace("-", "")
-    invoice_time = invoice["issue_time"].replace(":", "")
-    invoice_number = invoice["id"]
-    file_name = vat_registration_number + "_" + invoice_date + "T" + invoice_time + "_" + invoice_number
-    progressive_name = frappe.model.naming.make_autoname(file_name)
-    progressive_name = progressive_name[:len(file_name)]
-    return progressive_name + ".xml"
+# TODO: What is this?
+def generate_einvoice_xml_fielname(vat_registration_number: str, issue_date: str, issue_time: str, invoice_number: str):
+    file_name = vat_registration_number + "_" + issue_date + "T" + issue_time + "_" + invoice_number + ".xml"
+    return file_name