+++ CHANGELOG.md
@@ -14,3 +14,7 @@ to a section with the version name.
   * Add a method to generate XML regarding invoice type.
 * Create E-Invoicing-Sync page to run the sync batch.
   * Initiate the batch flow to Sync E-invoices Individually.
+* ZATCA Business Settings
+  * Onboarding: Compliance and production CSID support
+  * Signing and QR generation
+  * Invoice reporting and clearance support, although it currently fails with bad request